requirements.txt
hvac==2.1.0
python-dotenv==1.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=1.0.0
pylint>=2.17.0 